requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.6
django-filter==2.4.0
django-import-export==2.4.0
django-oauth-toolkit==1.3.3
django-stdimage==5.1.1
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
jdcal==1.4.1
Markdown==3.3.3
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==6.2.2
pkg-resources==0.0.0
progressbar2==3.53.1
psycopg2-binary==2.8.6
python-utils==2.4.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.26.2
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==3.4.0

pkg_resources==0.0.0
